refactor: move backup common types to backup_common.h
