+++ src/common/backup_common.cpp
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "backup_utils.h"
+#include "backup_common.h"
 #include "replica/backup/cold_backup_context.h"
 
 namespace dsn {
@@ -26,6 +26,16 @@ const std::string cold_backup_constant::CURRENT_CHECKPOINT("current_checkpoint")
 const std::string cold_backup_constant::BACKUP_METADATA("backup_metadata");
 const std::string cold_backup_constant::BACKUP_INFO("backup_info");
 const int32_t cold_backup_constant::PROGRESS_FINISHED = 1000;
+
+const std::string backup_restore_constant::FORCE_RESTORE("restore.force_restore");
+const std::string backup_restore_constant::BLOCK_SERVICE_PROVIDER("restore.block_service_provider");
+const std::string backup_restore_constant::CLUSTER_NAME("restore.cluster_name");
+const std::string backup_restore_constant::POLICY_NAME("restore.policy_name");
+const std::string backup_restore_constant::APP_NAME("restore.app_name");
+const std::string backup_restore_constant::APP_ID("restore.app_id");
+const std::string backup_restore_constant::BACKUP_ID("restore.backup_id");
+const std::string backup_restore_constant::SKIP_BAD_PARTITION("restore.skip_bad_partition");
+const std::string backup_restore_constant::RESTORE_PATH("restore.restore_path");
 
 namespace cold_backup {
