+++ src/common/replication_common.cpp
@@ -610,16 +610,6 @@ replication_options::check_if_in_black_list(const std::vector<std::string> &blac
     return false;
 }
 
-const std::string backup_restore_constant::FORCE_RESTORE("restore.force_restore");
-const std::string backup_restore_constant::BLOCK_SERVICE_PROVIDER("restore.block_service_provider");
-const std::string backup_restore_constant::CLUSTER_NAME("restore.cluster_name");
-const std::string backup_restore_constant::POLICY_NAME("restore.policy_name");
-const std::string backup_restore_constant::APP_NAME("restore.app_name");
-const std::string backup_restore_constant::APP_ID("restore.app_id");
-const std::string backup_restore_constant::BACKUP_ID("restore.backup_id");
-const std::string backup_restore_constant::SKIP_BAD_PARTITION("restore.skip_bad_partition");
-const std::string backup_restore_constant::RESTORE_PATH("restore.restore_path");
-
 const std::string replica_envs::DENY_CLIENT_WRITE("replica.deny_client_write");
 const std::string replica_envs::WRITE_QPS_THROTTLING("replica.write_throttling");
 const std::string replica_envs::WRITE_SIZE_THROTTLING("replica.write_throttling_by_size");
