+++ src/replica/backup/replica_backup_server.h
@@ -20,7 +20,7 @@
 #include <dsn/dist/replication/replication_types.h>
 #include <dsn/cpp/rpc_holder.h>
 
-#include "common/replication_common.h"
+#include "common/backup_common.h"
 
 namespace dsn {
 namespace replication {
