+++ src/replica/test/replica_test.cpp
@@ -21,10 +21,10 @@
 #include <gtest/gtest.h>
 #include "runtime/rpc/network.sim.h"
 
-#include "common/backup_utils.h"
+#include "common/backup_common.h"
 #include "replica_test_base.h"
 #include "replica/replica_http_service.h"
-#include "common/backup_utils.h"
+#include "common/backup_common.h"
 
 namespace dsn {
 namespace replication {
